Refactoring sobre el commit anterior
