+++ app/src/main/java/com/example/fran/imachineappv2/Classifier.java
@@ -5,7 +5,7 @@ import android.graphics.Bitmap;
 import java.util.List;
 
 /**
- * Created by amitshekhar on 17/03/18.
+ * Based on: https://github.com/amitshekhariitbhu/Android-TensorFlow-Lite-Example
  */
 
 public interface Classifier {
